Allow capturing screen from popup window
